+++ tests/testthat/test-networks.R
@@ -4,6 +4,6 @@ test_that("riem_networks returns the right output",{
   skip_on_cran()
   output <- riem_networks()
   expect_is(output, "tbl_df")
-  expect_is(output$code, "factor")
-  expect_is(output$name, "factor")
+  expect_is(output$code, "character")
+  expect_is(output$name, "character")
 })